requirements.txt
opencv-python
moviepy
Pillow